core/requirements.txt
pika
Pillow
json